+++ tempodb/encoding/vparquet/encoding.go
@@ -2,7 +2,9 @@ package vparquet
 
 import (
 	"context"
+	"time"
 
+	"github.com/google/uuid"
 	"github.com/grafana/tempo/pkg/model"
 	"github.com/grafana/tempo/tempodb/backend"
 	"github.com/grafana/tempo/tempodb/encoding/common"
@@ -30,4 +32,14 @@ func (v Encoding) CopyBlock(ctx context.Context, meta *backend.BlockMeta, from b
 
 func (v Encoding) CreateBlock(ctx context.Context, cfg *common.BlockConfig, meta *backend.BlockMeta, i common.Iterator, dec model.ObjectDecoder, r backend.Reader, to backend.Writer) (*backend.BlockMeta, error) {
 	return CreateBlock(ctx, cfg, meta, i, dec, r, to)
+}
+
+// OpenAppendBlock opens an existing appendable block
+func (v Encoding) OpenAppendBlock(filename string, path string, ingestionSlack time.Duration, additionalStartSlack time.Duration, fn common.RangeFunc) (common.AppendBlock, error, error) {
+	panic("jpe")
+}
+
+// CreateAppendBlock creates a new appendable block
+func (v Encoding) CreateAppendBlock(id uuid.UUID, tenantID string, filepath string, e backend.Encoding, dataEncoding string, ingestionSlack time.Duration) (common.AppendBlock, error) {
+	panic("jpe")
 }
